[fusion] Triggering dynamic slice fusion only for loops or constants
This patch adds a precondition to dynamic-slice-fusion with DUS
operations. It should only be generated when the offsets in DUS are
either loop iteration offsets or constant operations. Without this,
we end up waiting for a D2H transfer of the offset which is not ideal.

This pattern can be updated later to allow for more cases when we have
an optimal fusion execution for them.