Fix bug in post order traversal of computation instructions
While creating post order traversal, an instruction may have a user
outside the computation. This is the case when we are constructing
new instructions to store in replacements for cloning the computation
later. This user should be ignored. Added test for the same.

Because of this, functions like `ToString()`, and
`GetUniqueGteInstruction()` encounter errors. They rely on post-order
traversal to have all the instructions.